# docs/api/openapi-user.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskCreateInput": {
        "additionalProperties": false,
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/TaskFileInput"
            },
            "maxItems": 10,
            "title": "Files",
            "type": "array"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "prompt": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "TaskCreateInput",
        "type": "object"
      },
      "TaskFileInput": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "maxLength": 28000000,
            "title": "Content Base64",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "content_base64"
        ],
        "title": "TaskFileInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "serviceBearer": {
        "description": "管理员创建的服务凭据，或 /oauth/token 签发的 client_credentials 访问令牌。",
        "scheme": "bearer",
        "type": "http"
      },
      "umeko_auth": {
        "description": "网页登录后保存 Cookie；管理与用户 Cookie 相互独立。",
        "in": "cookie",
        "name": "umeko_auth",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Umeko 通用 Agent 基座的服务化 API。用户、Session 与对话历史保存在 SQLite；附件与产物保存在用户与 Session 隔离的服务端目录。",
    "title": "Umeko API",
    "version": "0.1.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/oauth/token": {
      "post": {
        "operationId": "oauth_token_oauth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string",
                    "writeOnly": true
                  },
                  "grant_type": {
                    "enum": [
                      "client_credentials"
                    ],
                    "type": "string"
                  },
                  "resource": {
                    "type": "string"
                  },
                  "scope": {
                    "type": "string"
                  }
                },
                "required": [
                  "grant_type"
                ],
                "type": "object"
              }
            }
          },
          "description": "client_credentials；也可用 HTTP Basic 传 client_id/client_secret。",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [],
        "summary": "Oauth Token",
        "tags": [
          "machine auth"
        ]
      }
    },
...
    "/v1/tasks": {
      "get": {
        "operationId": "list_tasks_v1_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Tasks V1 Tasks Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "machine tasks"
        ]
      },
      "post": {
        "operationId": "submit_task_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Task V1 Tasks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "Submit Task",
        "tags": [
          "machine tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Task V1 Tasks  Task Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "Get Task",
        "tags": [
          "machine tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "download_artifact_v1_tasks__task_id__artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "文件响应；实际 Content-Type 随文件类型变化，目录下载为 ZIP。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "Download Artifact",
        "tags": [
          "machine tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Task V1 Tasks  Task Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "Cancel Task",
        "tags": [
          "machine tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/events": {
      "get": {
        "operationId": "task_events_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "持久化任务事件 SSE；保留最近 500 条，支持 after 与 Last-Event-ID。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "serviceBearer": []
          },
          {
            "umeko_auth": []
          }
        ],
        "summary": "Task Events",
        "tags": [
          "machine tasks"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000"
    }
  ]
}